requirements.txt
attrs==16.2.0
cffi==1.9.1
constantly==15.1.0
cryptography==1.5.3
cssselect==1.0.0
Django==1.10.3
djangorestframework==3.5.3
idna==2.1
incremental==16.10.1
lxml==3.6.4
parsel==1.0.3
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==16.2.0
python-dateutil==2.6.0
queuelib==1.4.2
Scrapy==1.2.1
scrapy-djangoitem==1.1.1
service-identity==16.0.0
six==1.10.0
Twisted==16.5.0
w3lib==1.16.0
zope.interface==4.3.2
